requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.1.1
djangorestframework==3.16.0
filelock==3.16.1
ipynb==0.5.1
platformdirs==4.3.6
psycopg==3.2.6
sqlparse==0.5.1
typing_extensions==4.13.1
tzdata==2024.2
virtualenv==20.26.5
